feat: Add ProjectMWidget for visualization
This commit adds a new ProjectMWidget class, which is a QOpenGLWidget
that will be used to render the projectM visualization.

The MainWindow has been updated to use this new widget as its central
widget. The CMakeLists.txt has been updated to include the new files
and to link against the Qt6 OpenGL and OpenGLWidgets modules.